Move autocompleter footer into autocompleter
[#INT-1010]
